created components directory and move App jsx into it
