+++ client/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App.jsx';
+import App from './components/App.jsx';
 
 const container = document.getElementById('root');
 const root = ReactDOM.createRoot(container);
